migration:clean fix, db shared memory increase (#111)
* Fixing migration:clean

The command did not work sometimes because it attempted to remove
the literal db/migrations/*.js and was failing to find it.

	modified:   commands.json

* Increasing shared memory for db in docker-compose.yml

Required because otherwise heavier Postgres queries will keep
failing. Since it's just a tmpfs resize, there is no RAM overhead
from this unless the filesystem is actually used.

	modified:   docker-compose.yml